Move growthbook setup from main.tsx to App.tsx
